requirements.txt
streamlit>=1.26
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
PyPDF2>=3.0.0
numpy>=1.24